[FX] Move some heavily used passes out of experimental
[ghstack-poisoned]
